Rename status to qmgr
