+++ metricbeat/docs/modules/ibmmq.asciidoc
@@ -13,7 +13,7 @@ This module periodically fetches metrics from a containerized distribution of IB
 [float]
 === Compatibility
 
-The ibmmq `status` metricset is compatible with a containerized distribution of IBM MQ (since version 9.1.0).
+The ibmmq `qmgr` metricset is compatible with a containerized distribution of IBM MQ (since version 9.1.0).
 The Docker image starts the `runmqserver` process, which spawns the HTTP server exposing metrics in Prometheus
 format ([source code](https://github.com/ibm-messaging/mq-container/blob/9.1.0/internal/metrics/metrics.go)).
 
@@ -36,7 +36,7 @@ in <<configuration-metricbeat>>. Here is an example configuration:
 ----
 metricbeat.modules:
 - module: ibmmq
-  metricsets: ['status']
+  metricsets: ['qmgr']
   period: 10s
   hosts: ['localhost:9157']
 
@@ -69,7 +69,7 @@ It also supports the options described in <<module-http-config-options>>.
 
 The following metricsets are available:
 
-* <<metricbeat-metricset-ibmmq-status,status>>
+* <<metricbeat-metricset-ibmmq-qmgr,qmgr>>
 
-include::ibmmq/status.asciidoc[]
+include::ibmmq/qmgr.asciidoc[]
