+++ metricbeat/docs/modules/ibmmq/qmgr.asciidoc
@@ -2,12 +2,12 @@
 This file is generated! See scripts/mage/docs_collector.go
 ////
 
-[[metricbeat-metricset-ibmmq-status]]
-=== IBM MQ status metricset
+[[metricbeat-metricset-ibmmq-qmgr]]
+=== IBM MQ qmgr metricset
 
 beta[]
 
-include::../../../../x-pack/metricbeat/module/ibmmq/status/_meta/docs.asciidoc[]
+include::../../../../x-pack/metricbeat/module/ibmmq/qmgr/_meta/docs.asciidoc[]
 
 This is a default metricset. If the host module is unconfigured, this metricset is enabled by default.
 
@@ -20,5 +20,5 @@ Here is an example document generated by this metricset:
 
 [source,json]
 ----
-include::../../../../x-pack/metricbeat/module/ibmmq/status/_meta/data.json[]
+include::../../../../x-pack/metricbeat/module/ibmmq/qmgr/_meta/data.json[]
 ----
