+++ x-pack/metricbeat/metricbeat.reference.yml
@@ -505,7 +505,7 @@ metricbeat.modules:
 
 #-------------------------------- IBM MQ Module --------------------------------
 - module: ibmmq
-  metricsets: ['status']
+  metricsets: ['qmgr']
   period: 10s
   hosts: ['localhost:9157']
