+++ x-pack/metricbeat/module/ibmmq/module.yml
@@ -1,3 +1,3 @@
 name: ibmmq
 metricsets:
-- status
+- qmgr
