+++ x-pack/metricbeat/module/ibmmq/qmgr/_meta/docs.asciidoc
@@ -1,0 +1,3 @@
+This is the `qmgr` metricset of the IBM MQ module. It collects status information for the Queue Manager.
+The manager is a system program that is responsible for maintaining the queues and ensuring that the messages
+in the queues reach their destination.
