+++ x-pack/metricbeat/module/ibmmq/qmgr/_meta/testdata/ibmmq-status.9.1.4.0-r1-amd64.plain-expected.json
@@ -1,12 +1,12 @@
 [
     {
         "event": {
-            "dataset": "ibmmq.status",
+            "dataset": "ibmmq.qmgr",
             "duration": 115000,
             "module": "ibmmq"
         },
         "metricset": {
-            "name": "status",
+            "name": "qmgr",
             "period": 10000
         },
         "prometheus": {
