+++ x-pack/metricbeat/module/ibmmq/qmgr/qmgr_test.go
@@ -2,17 +2,15 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-// +build integration
+// +build !integration
 
-package stats
+package qmgr
 
 import (
 	"os"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
-
-	"github.com/elastic/beats/libbeat/tests/compose"
+	"github.com/elastic/beats/libbeat/logp"
 	"github.com/elastic/beats/metricbeat/mb"
 	mbtest "github.com/elastic/beats/metricbeat/mb/testing"
 
@@ -27,22 +25,8 @@ func init() {
 	mb.Registry.SetSecondarySource(mb.NewLightModulesSource("../../../module"))
 }
 
-func TestFetch(t *testing.T) {
-	service := compose.EnsureUp(t, "ibmmq")
+func TestEventMapping(t *testing.T) {
+	logp.TestingSetup()
 
-	f := mbtest.NewFetcher(t, getConfig(service.Host()))
-	events, errs := f.FetchEvents()
-	if len(errs) > 0 {
-		t.Fatalf("Expected 0 error, had %d. %v\n", len(errs), errs)
-	}
-	assert.NotEmpty(t, events)
-	t.Logf("%s/%s event: %+v", f.Module().Name(), f.Name(), events[0])
-}
-
-func getConfig(host string) map[string]interface{} {
-	return map[string]interface{}{
-		"module":     "ibmmq",
-		"metricsets": []string{"status"},
-		"hosts":      []string{host},
-	}
+	mbtest.TestDataFiles(t, "ibmmq", "qmgr")
 }
