--- x-pack/metricbeat/module/ibmmq/status/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the stats metricset of the IBM MQ module.
